gopls/matcher: don't give bonus to case insensitive match
Previously the user given pattern "A" would match "Ant" and "ant"
equally well. Now it will match "Ant" better than "ant".

There already was a case sensitive match bonus, but it was combined
with a "head" character bonus, so "Ant" got the case sensitive bonus,
and "ant" got the head match bonus (and they tied).

Tweak the head match bonus to not apply if the pattern is case
sensitive. In other words, if the user is explicitly putting capital
letters in the pattern, don't give a bonus if the pattern char's case
differs from that of the candidate char.

For example, if the pattern is "Foo", the "F" no longer gets a bonus
for matching the "f" in "fooBar".

I tweaked the test to test strict relative ranking. Previously it
allowed adjacent test inputs to have the same score, but I wanted to
test that "N" matches "Name" strictly better than "name".

Fixes golang/go#64224

Change-Id: Ibce21985d5c904167cec239d62a4f53e814a9b08
Reviewed-on: https://go-review.googlesource.com/c/tools/+/592519
Reviewed-by: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>