gopls/doc/release: finalize v0.16.0 release notes
The images are intentionally missing; we don't want to
bloat the repo with PNG files, so I plan to set up
a process for storing images in a blob store later
today, at which point I will update the URLs to point
there. But we can proceed with the release before that.

(Many of these images will be shared with the gopls durable
documentation slowly coming together in CL 583316.)

Change-Id: I508b13ad2ae3b77f08dfeb895472db09542661b2
Reviewed-on: https://go-review.googlesource.com/c/tools/+/593476
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>