go/analysis/passes/buildtag: retire Go 1.15 support
Go 1.17 is long since out.

Furthermore, CL 516575 added a use of parser.SkipObjectResolution to
a file with a go1.16 build constraint. This used to be fine, but not
now that vet checks for this (see go.dev/issue/66092) since that API
is new to Go 1.17.

For golang/go#66092.

Change-Id: I749b2aa52e02308415c27028be15a436ee11d95c
Reviewed-on: https://go-review.googlesource.com/c/tools/+/593297
Reviewed-by: Michael Matloob <matloob@golang.org>
Reviewed-by: Sam Thanawalla <samthanawalla@google.com>
Auto-Submit: Dmitri Shuralyov <dmitshur@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>