[gopls-release-branch.0.16] go/analysis/passes/printf: add missing Unalias call
The maybePrintfWrapper function checks to see that a function
has the form of a printf wrapper, but it wrongly assumed that
the representation of the type of the "args ...any" parameter
is exactly interface{}, not a named alias. This will not work
with gotypesalias=1.

Unfortunately our CL system failed to report this (or indeed
any gotypesalias=1 coverage at all) because of a bug in the
Go bootstrapping process that, in the absence of a go.work file
(which sets the language version to go1.23), the default values
of the GODEBUG table were based on an older version of Go.
(The problem was only noticed when running a test of unitchecker
locally in the context of issue 68796.)

Also, the problem wasn't caught by our existing tests of the
printf checker because they all pre-date "any", and so spelled
it "interface{}".

This CL will need to be vendored into the go1.23 release.

Updates golang/go#68744
Updates golang/go#68796

Change-Id: I834ea20c2a684ffcd7ce9494d3700371ae6ab3c1
Reviewed-on: https://go-review.googlesource.com/c/tools/+/603938
Auto-Submit: Alan Donovan <adonovan@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
(cherry picked from commit 28ba9914c6b79f6cf3a56cc477398f7fd686c84d)
Reviewed-on: https://go-review.googlesource.com/c/tools/+/609435
TryBot-Bypass: Robert Findley <rfindley@google.com>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>