go/callgraph/static: avoid ssautil.AllFunctions
AllFunctions is complex mess. This change rewrites the static
algorithm to avoid it. It does reduce the number of
call graph nodes that are discovered (by something like 15%),
and the running time by a similar amount, but the principle
is slightly more defensible.

Also, document exactly what it does, and why it is the way
it is.

Updates golang/go#69231

Change-Id: I7e25237f0908315602ba0092083f247a140b9e22
Reviewed-on: https://go-review.googlesource.com/c/tools/+/609280
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>