internal/modindex: package for indexing GOMODCACHE
This CL contains the first part of a package for maintaining an
on-disk index of the module cache. The index is stored as text.
Eventually it will consist of a header, followed by groups of lines,
one for each import path, and sorted by package name. The groups of lines
start with a header containing the package name, import path, name of the
directory, and semantic version,
followed (but not in this first CL) by lines, each of which contains
information about one exported symbol.

This CL only contains the code for computing and updating the information
about directories and import paths, and reading the index. It does not
compute anything about exported symbols, which will be in the next CL,
and hence it does not present an API for looking up information about
completion of selectors.

There is a test that among directories with the same import path it can
find the one with the largest semantic version.

Change-Id: I0883ea732cf34f6700f5495e6dfd594e8f286af9
Reviewed-on: https://go-review.googlesource.com/c/tools/+/612355
TryBot-Bypass: Peter Weinberger <pjw@google.com>
Reviewed-by: Robert Findley <rfindley@google.com>