go/ssa: add test utility for building a single file
Adds a new testing utility function that builds a module an
package from the contents of a single file. This allows for
significant simplifications of the ssa tests. One benefit is
removing the use of the deprecated loader package.

Refactors loadPackages to work from an fs.FS.

For golang/go#69556.

Change-Id: If25f103982c82ea7945339e50f6c0e8576d276fa
Reviewed-on: https://go-review.googlesource.com/c/tools/+/614735
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>