gopls/internal/telemetry/cmd/stacks: fix bad tmpreaper interaction
This CL causes stacks to check for the presence of a go.mod file
for each cloned repo, not just a directory, as the tmpreaper (at
least on macOS) deletes the stales files before it deletes the
directories. If the file is not found, we delete the directory
tree.

Change-Id: Idf4990f707704e12ae15120269cf7d490e406b41
Reviewed-on: https://go-review.googlesource.com/c/tools/+/620035
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Auto-Submit: Alan Donovan <adonovan@google.com>