go/analysis/passes/unusedwrite: silence if unsafe is imported
As reported in golang/go#67684, the unusedwrite analyzer may have false
positives when used with unsafe, so silence it if unsafe is imported.

This is a rather large hammer, but there is a high likelihood of such
false positives, and they significantly detracting from the usefulness
of the analysis. Specifically: this noise causes Go developers to
disable the analyzer.

Fixes golang/go#67684

Change-Id: I855a5da6124a0495c7eb11722b466824689780a3
Reviewed-on: https://go-review.googlesource.com/c/tools/+/621817
Auto-Submit: Robert Findley <rfindley@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>