gopls/internal/template: fix completion token boundary conditions
The template completion logic checked to see if the cursor position was
inside a logical {{ }} block, but the boundary conditions in the
relevant checks was wrong: in one case, a panic would occur if the
cursor was in between the pair of left braces. In another, it would
panic even if the cursor was at the start of a pair of left braces.

Fix the boundary conditions, with a test.

Fixes golang/go#57621

Change-Id: I826349906ee1ae67b2c5378e1b59d56e94c14fb2
Reviewed-on: https://go-review.googlesource.com/c/tools/+/622875
Auto-Submit: Robert Findley <rfindley@google.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>