go/analysis/passes/copylock: test for noCopy for sync Map, Mutex, Once
Right now the copylock tests check for details on these types that are
implementation details and/or have some other problem (a less useful
error message, or a false negative altogether).

CL 627777 modifies the sync package to attach explicit anonymous noCopy
fields on each of these types. This change updates the tests to match
that, which also helps with and/or resolves a couple issues with
copylock captured in the test suite.

Note: this change also temporarily disables a couple of the problematic
tests. In the next CL we'll re-enable them, once CL 627777 lands.

Change-Id: I100c71ea05b4f08595e37d0c8e81f9543abe7d74
Reviewed-on: https://go-review.googlesource.com/c/tools/+/628435
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
Reviewed-by: Tim King <taking@google.com>