internal/refactor/inline: fix spurious caller mutation
A failing test in a later CL was rooted in a bug allowing the inliner to
mutate the caller syntax: during literalization we were clearing
positions after inserting the binding decl, and the binding decl may
refer to syntax in the caller.

This was not caught by tests, because the deepHash utility itself had a
bug: binary.Write is a no op for ints or uints, because they are not
fixed width (this was rather surprising behavior). This is fixed by
adding handling for Int and Uint values, and delegating to binary.Write
only in the cases that are known to be supported.

Change-Id: Icdf03f4d2cbc91b347b876cc39d8a48484c90c7c
Reviewed-on: https://go-review.googlesource.com/c/tools/+/631295
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>