gopls/internal/golang: simplify package name collection in add test
Previously, package imports were queried by ID to create a
comprehensive map of package paths and their local names. This
approach is not necessary as the "types.Package" is readily available
when qualifiers are called.

The import maps now use the package path as the key and the local
rename as the value. An empty value means that the package is imported
without a local rename.

Change-Id: I0d0039217dc335b104c91a6bf541f0bfa4d3c509
Reviewed-on: https://go-review.googlesource.com/c/tools/+/631996
Auto-Submit: Hongxiang Jiang <hxjiang@golang.org>
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>