gopls/internal/cache: move symbols to the file cache
Prior to this CL, gopls would parse all files on the first call to
workspace symbols, to extract symbol information. This results in a slow
initial request to workspace symbols, and a significant amount of peak
memory, until the parseCache evicts the resulting parsed files.

This CL replaces that on-demand parsing with a new  operation to collect
and persistently store symbols for a given package's files. This
operation will also be useful for other gopls functionality, such as
unimported completion or goimports.

For the most part, the existing logic to collect symbols from a file is
unchanged, with the notable exception that it is made an infallible
operation: the collector simply skips a symbol when its range is
invalid. In the future this should be a bug report, but we cannot
currently enforce this due to inconsistencies in go/parser.

Additionally, the following minor improvements are included:
- The logic for downranking non-workspace packages is updated to use
  package metadata. This required exposing the immutable
  WorkspacePackages map.
- The redundant symbolInformation struct is removed.
- Benchmarks are updated to avoid external device errors when copying
  profiles from temp directories.

Change-Id: I8bfcd8aa98663f8531ba56341676cb19d941d655
Reviewed-on: https://go-review.googlesource.com/c/tools/+/634796
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Alan Donovan <adonovan@google.com>