internal/gcimporter: remove end-of-section marker from size
Remove the bytes for the end-of-section marker "\n$$\n" from
the end of the exportdata by descreasing the returned size
appropriately. This marker is not a part of the export data
so it should be removed. Updates the unified IR reader to not
expect the marker to be present.

Updates golang/go#70651

Change-Id: Id3324ff21162d80f8e3d1c9792e9157b01ae1a3f
Reviewed-on: https://go-review.googlesource.com/c/tools/+/633296
Reviewed-by: Robert Findley <rfindley@google.com>
Commit-Queue: Tim King <taking@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>