gopls/internal/golang/stubmethods: refine crash into bug report
This CL refine an as-yet unreproducible crash into a bug report. It
seems clear that the code intended to check that sig != nil, yet instead
checked that sig.Results != nil. With that said, I don't think it's
possible to have a return statement without an enclosing signature. For
now, turn this into a bug report.

For golang/go#70666

Change-Id: I79031d437d643f6d70360e3abde86166176647b3
Reviewed-on: https://go-review.googlesource.com/c/tools/+/633375
Reviewed-by: Alan Donovan <adonovan@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>