internal/refactor/inline/analyzer: use "//go:fix inline" annotation
This CL changes the proof-of-concept "inliner" analyzer to
use "//go:fix inline" as the marker of a function that should
be inlined. This seems to be the syntax emerging from the
proposal golang/go#32816.

Also, skip inlinings that cause literalization of the callee.
Users of batch tools only want to see reductions.

Updates golang/go#32816
Updates golang/go#70717

Change-Id: I6d70118f69b7c35f51e1d51863a0312b5dcc3b63
Reviewed-on: https://go-review.googlesource.com/c/tools/+/634919
Reviewed-by: Robert Findley <rfindley@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>