gopls/internal/analysis/modernize: quick fixes to modernize Go code
This CL is the start of an analyzer that offers a suite of code
transformations to modernize Go code, by taking advantage of
newer features of the language when it would clarify the code
to do so. Because it reports diagnostics on correct code,
its severity is downgraded to INFO.

Only one is implemented so far: the replacement of if/else
conditional assignments by a call to min or max.

+ relnote and test

Updates golang/go#70815

Change-Id: I686db57c2b95dfa399b44e5c6fdc478272cee084
Reviewed-on: https://go-review.googlesource.com/c/tools/+/635777
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Robert Findley <rfindley@google.com>